requirements.txt
fastapi
uvicorn
openai
python-dotenv
beutifulsoup4
